added css changes
