+++ src/main/java/com/khadri/jakarta/stock/form/BackCoverForm.java
@@ -1,31 +1,35 @@
-package com.khadri.stock.mobile.form;
+package com.khadri.jakarta.stock.form;
 
 import java.sql.Date;
 
-public class HeadSetForm {
+public class BackCoverForm {
 	private String productBrand;
 	private Double productPrice;
 	private String productModel;
 	private Date arrivedDateTime;
-	public HeadSetForm(String productBrand, Double productPrice, String productModel, Date arrivedDateTime) {
+
+	public BackCoverForm(String productBrand, Double productPrice, String productModel, Date arrivedDateTime) {
 		super();
 		this.productBrand = productBrand;
 		this.productPrice = productPrice;
 		this.productModel = productModel;
 		this.arrivedDateTime = arrivedDateTime;
 	}
+
 	public String getProductBrand() {
 		return productBrand;
 	}
+
 	public Double getProductPrice() {
 		return productPrice;
 	}
+
 	public String getProductModel() {
 		return productModel;
 	}
+
 	public Date getArrivedDateTime() {
 		return arrivedDateTime;
 	}
 
-	
 }
