+++ src/main/java/com/khadri/jakarta/stock/form/HeadSetForm.java
@@ -1,35 +1,31 @@
-package com.khadri.stock.mobile.form;
+package com.khadri.jakarta.stock.form;
 
 import java.sql.Date;
 
-public class MobileForm {
+public class HeadSetForm {
 	private String productBrand;
 	private Double productPrice;
 	private String productModel;
 	private Date arrivedDateTime;
-
-	public MobileForm(String productBrand, Double productPrice, String productModel, Date arrivedDateTime) {
+	public HeadSetForm(String productBrand, Double productPrice, String productModel, Date arrivedDateTime) {
 		super();
 		this.productBrand = productBrand;
 		this.productPrice = productPrice;
 		this.productModel = productModel;
 		this.arrivedDateTime = arrivedDateTime;
 	}
-
 	public String getProductBrand() {
 		return productBrand;
 	}
-
 	public Double getProductPrice() {
 		return productPrice;
 	}
-
 	public String getProductModel() {
 		return productModel;
 	}
-
 	public Date getArrivedDateTime() {
 		return arrivedDateTime;
 	}
 
+	
 }
