+++ src/main/webapp/WEB-INF/web.xml
@@ -9,7 +9,7 @@
 	</welcome-file-list>
 	<servlet>
 		<servlet-name>sap</servlet-name>
-		<servlet-class>com.khadri.stock.mobile.servlet.StockAddPageServlet</servlet-class>
+		<servlet-class>com.khadri.jakarta.stock.servlet.add.StockAddPageServlet</servlet-class>
 	</servlet>
 	<servlet-mapping>
 		<servlet-name>sap</servlet-name>
@@ -17,7 +17,7 @@
 	</servlet-mapping>
 	<servlet>
 		<servlet-name>stockadd1</servlet-name>
-		<servlet-class>com.khadri.stock.mobile.servlet.StockAddServlet</servlet-class>
+		<servlet-class>com.khadri.jakarta.stock.servlet.add.StockAddServlet</servlet-class>
 	</servlet>
 	<servlet-mapping>
 		<servlet-name>stockadd1</servlet-name>
