+++ src/main/webapp/add_product.html
@@ -1,9 +1,9 @@
 
 <!DOCTYPE html>
 <html>
-<body>
+<body bgcolor="#eee">
 	<form action="add" method="post">
-		<table border="3">
+		<table border="3" >
 			<thead>
 				<h2>AddMobileProduct</h2>
 			</thead>
