--- src/main/webapp/bottom_left_stock.html
@@ -1,28 +1,0 @@
-<html>
-<body>
-	<table>
-		<thead>
-			<h2>Stock details</h2>
-		</thead>
-		<tbody>
-			<tr>
-				<td><a href="stockaddpage" target="bottom_right">add</a></td>
-			</tr>
-			<tr>
-				<td><a href="bottom_right_stock_modify.html"
-					target="bottom_right">modify</a></td>
-			</tr>
-			<tr>
-				<td><a href="bottom_right_stock_view.html" target="bottom_right">view</a></td>
-			</tr>
-			<tr>
-				<td><a href="bottom_right_stock_viewall.html"
-					target="bottom_right">viewAll</a></td>
-			</tr>
-			<tr>
-				<td><a href="bottom_right_stock_delete.html" target="bottom_right">delete</a></td>
-			</tr>
-		</tbody>
-	</table>
-</body>
-</html>
