+++ src/main/webapp/top.html
@@ -1,9 +1,9 @@
 <html>
-<body>
+<body bgcolor="#eee">
 	<div align="center">
 		<table>
 
-			<h1 style="color: #bd3d2f;">Mobile Shop App</h1>
+			<h1 style="color: #04AA6D;">Mobile Shop App</h1>
 
 		</table>
 	</div>
